Add files via upload
A movie website that includes posters of movies with their information and comments. Comments are posted as voices then they will convert to text. If the comment is detected as violent, an alert will be shown and it will not be displayed. Comments can also be displayed in three languages.